Seperate data collection and plotting
After this change, the mixin only contains the functionality required
to collect all the relevant data for plotting. The functions that actually
create the plots only depend on the new plot_data_and_config dict. This
allows us to store plot data and options and create plots later, as well
as possibly creating plots that depend on more than one optimization run.