Make order of PlotMixin/GoalGeneratorMixin irrelevant
Previously, if the PlotMixin was initialized before the GoalGeneratorMixin, the goals were not loaded yet and as a result the goal_generator plots were missing. Now, the loading of the goals is moved to a seperate mixin. The function is called only once, by the first of plot_mixin or goal_generator_mixin.